+++ src/HWGT/comparators/PersonComparatorByBirth.java
@@ -1,12 +1,11 @@
 package HWGT.comparators;
 
-import HWGT.Person;
+import HWGT.model.Basic;
+
 
 import java.util.Comparator;
 
-public class PersonComparatorByBirth implements Comparator<Person> {
+public class PersonComparatorByBirth<T extends Basic> implements Comparator<T> {
     @Override
-    public int compare(Person o1, Person o2) {
-        return o1.getDateOfBirth().compareTo(o2.getDateOfBirth());
-    }
+    public int compare(T o1, T o2) { return o1.getDateOfBirth().compareTo(o2.getDateOfBirth()); }
 }
