+++ src/HWGT/comparators/PersonComparatorByName.java
@@ -1,13 +1,12 @@
 package HWGT.comparators;
 
-import HWGT.Person;
+import HWGT.model.Basic;
+
 
 import java.util.Comparator;
 
-public class PersonComparatorByName implements Comparator<Person> {
+public class PersonComparatorByName<T extends Basic> implements Comparator<T> {
     @Override
-    public int compare(Person o1, Person o2) {
-        return o1.getName().compareTo(o2.getName());
-    }
+    public int compare(T o1, T o2) {return o1.getName().compareTo(o2.getName());}
 
 }
