+++ src/HWGT/model/Basic.java
@@ -1,0 +1,28 @@
+package HWGT.model;
+
+public class Basic {
+    public String getName(){
+        return null;
+    }
+    public void addChild(String string){
+
+    }
+    public void setDateOfBirth(Integer birthDate){
+
+    }
+    public void setDateOfDeath(Integer deathDate){
+
+    }
+    public void setFather(String fatherName){
+
+    }
+    public void setMother(String motherName){
+
+    }
+    public Integer getDateOfBirth(){
+        return null;
+    }
+    public Integer getDateOfDeath(){
+        return null;
+    }
+}
