+++ src/HWGT/model/Repo.java
@@ -1,8 +1,8 @@
-package HWGT;
+package HWGT.model;
 
 import java.io.*;
 
-public class Repo implements FileData{
+public class Repo implements FileData {
 
     @Override
     public void save(Serializable serializable, String filename) throws IOException {
